ecs: port petal ring kinematics, repair two player sync-window bugs, add p99 gate
Checkpoint of an interrupted multi-agent run. Repair and Petals completed with
two reviews each; the bots stage was killed mid-implementation, so
botManager.ts carries partial, UNREVIEWED work (+272/-121).

REPAIR (reviewed)
- PVP knockback was silently discarded: applyPvpDamage wrote victim.x from
  inside the ATTACKER's updatePlayerState, after the victim's move had already
  been integrated, so the ECS overwrote it.
- A yggdrasil revive set otherPlayer.isDead from another player's update, same
  class of bug.
Both are the write-outside-the-sync-window failure this rewrite keeps
reproducing.

PETALS (reviewed: one lens ship, one fix-first)
Petal RING KINEMATICS move to the ECS (ecs/systems/petalRing.ts + the PetalRing
component): orbit-slot layout, orbit point, the spring/glide integrator and the
per-instance state behind it. Per-instance health, the break/cooldown/reload
machine, petal combat, fields/auras and the action VM stay legacy.

It is stepped ONE INSTANCE AT A TIME from the legacy petal loop rather than as
a scheduled system, because that loop interleaves kinematics with effects and
instance k's effects change instance k+1's kinematics. There is no petal sync
at all: the ECS is the sole writer of petal kinematic state, so unlike mobs and
players there is no window and no field two sides can fight over.

A reviewer proved the ownership rather than arguing it: petalOrbitPhase has
exactly two writers and one reader, the legacy mirror is write-only, and both
sync paths touch PlayerModifiers only with per-field world.set.

playerModifiers deliberately STOPPED integrating petalOrbitPhase. That system
is disabled during the cutover, so integrating only there would freeze every
orbit — and integrating in both places once it is enabled would silently double
the rate. Its multiplicative stacking also disagrees with the legacy additive
stacking, so it cannot drive the ring until that is reconciled.

HARNESS (fixed here, after both reviewers rejected the first attempt)
The harness now reports p50/p95/p99 and ASSERTS on p99. The implementing agent
had also asserted on maxTickMs; both reviewers correctly called that out — the
max is exactly the quantity that spikes, so asserting on it turns a required
gate into a coin flip, and a gate that fails at random trains everyone to
ignore it. Max is now reported loudly with guidance for attributing a spike
(one system elevated = algorithmic; several = a process-wide stall) but never
fails the build. This closes the gap where the harness printed "OK" through
63.9ms and 107ms spikes.

Current: mean 2.2ms, p99 5.2ms, worst 33.9ms with the note firing correctly.

OUTSTANDING
- botManager.ts partial and unreviewed; the bots stage needs re-running.
- The petal test oracle never asserts PetalKinematicsResult.angle. A reviewer
  demonstrated the gap by fault injection: corrupting angle leaves the oracle
  green while four other injected faults are caught. Unproven, not wrong.

Gates: both tsconfigs clean, test:ecs passes, harness passes on mean and p99.

+++ src/ecs/components/player.ts
@@ -168,8 +168,21 @@ export const PlayerModifiers = defineComponent('PlayerModifiers', {
      * Continuous integral of the rotation-speed modifier. Kept as an
      * accumulated phase so that changing rotation speed mid-flight bends the
      * orbit rate instead of snapping the angle.
+     *
+     * f64, and that is not a spare-bytes decision. This is an UNBOUNDED
+     * accumulator — it grows by `rotationSpeed * deltaTime` every tick for the
+     * whole of a session and is never wrapped, so a long-lived flower reaches
+     * ~1e5 after a day of play. It is then multiplied by up to the petal's speed
+     * multiplier and fed to cos/sin, and the result is the petal position that
+     * gets hashed for the broadcast change detector. In f32 the representable
+     * step at 1e5 is ~0.008 while a tick's increment is ~0.03, so the phase
+     * would quantise to a coarse ladder: the whole ring would visibly step
+     * rather than rotate, and — because the hash would change on a different
+     * schedule from the legacy value — the broadcast would churn. The legacy
+     * `ServerPlayer.petalOrbitPhase` this replaces was a plain JS double, and
+     * `ecs/bench/petal_cutover_check.ts` asserts exact equality against it.
      */
-    petalOrbitPhase: 'f32',
+    petalOrbitPhase: 'f64',
 });
 
 /** Server-computed absolute petal positions, broadcast for accurate rendering. */
